Implement disk led rendering on Forge
