+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/render/CubeBuilder.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage2.platform.fabric.render;
+package com.refinedmods.refinedstorage2.platform.common.render;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
