+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/ClientModInitializer.java
@@ -1,16 +1,19 @@
 package com.refinedmods.refinedstorage2.platform.forge;
 
+import com.refinedmods.refinedstorage2.platform.common.content.BlockEntities;
 import com.refinedmods.refinedstorage2.platform.common.content.Blocks;
 import com.refinedmods.refinedstorage2.platform.common.content.Items;
 import com.refinedmods.refinedstorage2.platform.common.content.Menus;
 import com.refinedmods.refinedstorage2.platform.common.render.model.ControllerModelPredicateProvider;
 import com.refinedmods.refinedstorage2.platform.common.screen.ControllerScreen;
 import com.refinedmods.refinedstorage2.platform.common.screen.DiskDriveScreen;
+import com.refinedmods.refinedstorage2.platform.forge.render.entity.DiskDriveBlockEntityRendererImpl;
 import com.refinedmods.refinedstorage2.platform.forge.render.model.DiskDriveModelLoader;
 
 import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraft.client.renderer.ItemBlockRenderTypes;
 import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraftforge.client.event.ModelRegistryEvent;
 import net.minecraftforge.client.model.ModelLoaderRegistry;
@@ -28,16 +31,7 @@ public final class ClientModInitializer {
         e.enqueueWork(ClientModInitializer::setRenderLayers);
         e.enqueueWork(ClientModInitializer::registerModelPredicates);
         e.enqueueWork(ClientModInitializer::registerScreens);
-    }
-
-    private static void registerScreens() {
-        MenuScreens.register(Menus.INSTANCE.getController(), ControllerScreen::new);
-        MenuScreens.register(Menus.INSTANCE.getDiskDrive(), DiskDriveScreen::new);
-    }
-
-    @SubscribeEvent
-    public static void onRegisterModels(ModelRegistryEvent e) {
-        ModelLoaderRegistry.registerLoader(createIdentifier("disk_drive"), new DiskDriveModelLoader());
+        registerBlockEntityRenderer();
     }
 
     private static void setRenderLayers() {
@@ -50,5 +44,19 @@ public final class ClientModInitializer {
 
     private static void registerModelPredicates() {
         Items.INSTANCE.getControllers().forEach(controllerBlockItem -> ItemProperties.register(controllerBlockItem, createIdentifier("stored_in_controller"), new ControllerModelPredicateProvider()));
+    }
+
+    private static void registerScreens() {
+        MenuScreens.register(Menus.INSTANCE.getController(), ControllerScreen::new);
+        MenuScreens.register(Menus.INSTANCE.getDiskDrive(), DiskDriveScreen::new);
+    }
+
+    @SubscribeEvent
+    public static void onRegisterModels(ModelRegistryEvent e) {
+        ModelLoaderRegistry.registerLoader(createIdentifier("disk_drive"), new DiskDriveModelLoader());
+    }
+
+    private static void registerBlockEntityRenderer() {
+        BlockEntityRenderers.register(BlockEntities.INSTANCE.getDiskDrive(), ctx -> new DiskDriveBlockEntityRendererImpl<>());
     }
 }
